Set environment variables manually
The Emacs.app Coca compile does not pick up the right environment variables by default.
This change includes code to set the PATH and exec-path prior to el-get running, so that el-get has
access to the version control tools it needs.  Following that, we can use exec-path-from-shell to read
the MANPATH and PYTHONPATH from a login shell.  This is needed to work around Apple disabling the environment.plist
in OSX 10.8 (Mountain Lion).